requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-openai
langchain-community
langchain-text-splitters
langchain-chroma
chromadb
pypdf
requests
python-dotenv
